Remove jQuery
Now that jQuery is optional and no longer the default for Rails, this
means we can drop this dependency and serve a little bit lighter load to
all our pages.

Make use of the `addEventListener` function on elements (the mirror of
`dispatchEvent`) to listen for the `ajax:success` event on our
`new_point_creator` form. The argument to the callback changed: the
second param (`data`) is now part of the first argument (`event`),
accessed as `event.detail[0]`.

Use the `Rails.ajax` function to get one of the niceties jQuery
provides: a wrapper around XHR inconsistencies. It is slightly different
from the jQuery implementation: it requires the HTTP verb (`type`), and
`success` is a function where we must manually dispatch over the HTTP
response code (`xhr.status`).